[DF-9698] Fixed export of cubemaps
Former-commit-id: c4f622efd620076c0a85cc3b08ab8a53a74e6637

+++ Sources/Internal/Render/Texture.cpp
@@ -509,7 +509,7 @@ bool Texture::LoadImages(eGPUFamily gpu, Vector<Image*>* images)
                 ImageSystem::Instance()->Load(imagePathnames[i], *images, params);
             }
 
-            params.baseMipmap = Max(baseMipMap - singleMipCount, 0u);
+            params.baseMipmap = Max((int32)baseMipMap - (int32)singleMipCount, 0);
             params.firstMipmapIndex = static_cast<uint32>(images->size());
             ImageSystem::Instance()->Load(imagePathnames[singleMipCount], *images, params);
         }
